Move KeyboardShortcutEditor to its own component
